[#3] move: ファイルの移動とAPIの呼び出し処理を追加
